backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# PDF Processing
PyMuPDF==1.23.8
pdfminer.six==20221105
pypdf==3.17.1

# NLP & Question Generation
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
sentence-transformers==2.2.2

# Document Export
reportlab==4.0.7
python-docx==1.1.0
weasyprint==60.1

# Utilities
aiofiles==23.2.1
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
